requirements.txt
certifi==2018.8.24
chardet==3.0.4
idna==2.7
pkg-resources==0.0.0
pyserial==3.4
requests==2.19.1
urllib3==1.23
